Refactor test code, rename test file.
